+++ frontend/src/api/solve.ts
